+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/components/selectedmessage/SelectedMessageMenu.kt
@@ -111,6 +111,7 @@ private fun SelectedMessageMenuPreview() {
         val messageOptionsStateList = defaultMessageOptionsState(
             selectedMessage = Message(),
             currentUser = User(),
+            isInThread = false,
             ownCapabilities = ChannelCapabilities.toSet(),
         )