+++ stream-chat-android-compose/src/test/kotlin/io/getstream/chat/android/compose/ui/messages/SelectedMessageMenuTest.kt
@@ -95,6 +95,7 @@ internal class SelectedMessageMenuTest : SnapshotTest {
             val messageOptions = defaultMessageOptionsState(
                 selectedMessage = selectedMessage,
                 currentUser = currentUser,
+                isInThread = false,
                 ownCapabilities = channelCapabilities,
             )