+++ stream-chat-android-compose/src/test/kotlin/io/getstream/chat/android/compose/util/extensions/MessageOptionItemVisibilityTest.kt
@@ -50,11 +50,12 @@ internal class MessageOptionItemVisibilityTest {
     @MethodSource("canThreadReplyToMessageArguments")
     fun `Verify canThreadReplyToMessage() extension function return proper value`(
         messageOptionItemVisibility: MessageOptionItemVisibility,
+        isInThread: Boolean,
         message: Message,
         ownCapabilities: Set<String>,
         expectedResult: Boolean,
     ) {
-        messageOptionItemVisibility.canThreadReplyToMessage(message, ownCapabilities) `should be` expectedResult
+        messageOptionItemVisibility.canThreadReplyToMessage(isInThread, message, ownCapabilities) `should be` expectedResult
     }
 
     @ParameterizedTest
@@ -404,26 +405,37 @@ internal class MessageOptionItemVisibilityTest {
         fun canThreadReplyToMessageArguments() = listOf(
             Arguments.of(
                 MessageOptionItemVisibility(isThreadReplyVisible = false),
+                randomBoolean(),
                 randomMessage(),
                 randomChannelCapabilities(),
                 false,
             ),
             Arguments.of(
                 MessageOptionItemVisibility(),
+                randomBoolean(),
                 randomMessage(syncStatus = randomSyncStatus(exclude = listOf(SyncStatus.COMPLETED))),
                 randomChannelCapabilities(),
                 false,
             ),
             Arguments.of(
                 MessageOptionItemVisibility(),
+                randomBoolean(),
                 randomMessage(),
-                randomChannelCapabilities(exclude = setOf(ChannelCapabilities.QUOTE_MESSAGE)),
+                randomChannelCapabilities(exclude = setOf(ChannelCapabilities.SEND_REPLY)),
                 false,
             ),
             Arguments.of(
                 MessageOptionItemVisibility(isThreadReplyVisible = true),
+                true,
                 randomMessage(syncStatus = SyncStatus.COMPLETED),
-                randomChannelCapabilities(include = setOf(ChannelCapabilities.QUOTE_MESSAGE)),
+                randomChannelCapabilities(include = setOf(ChannelCapabilities.SEND_REPLY)),
+                false,
+            ),
+            Arguments.of(
+                MessageOptionItemVisibility(isThreadReplyVisible = true),
+                false,
+                randomMessage(syncStatus = SyncStatus.COMPLETED),
+                randomChannelCapabilities(include = setOf(ChannelCapabilities.SEND_REPLY)),
                 true,
             ),
         )