+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/feature/messages/list/internal/MessageListViewExtensions.kt
@@ -42,10 +42,12 @@ internal fun MessageListViewStyle.canReplyToMessage(
 )
 
 internal fun MessageListViewStyle.canThreadReplyToMessage(
+    isInThread: Boolean,
     message: Message,
     ownCapabilities: Set<String>,
 ): Boolean = canThreadReplyToMessage(
     threadsEnabled = threadsEnabled,
+    isInThread = isInThread,
     message = message,
     ownCapabilities = ownCapabilities,
 )